+++ src/plone/jsonapi/routes/providers/plonesites.py
@@ -6,9 +6,9 @@ from plone.jsonapi.routes.api import get_batched
 
 
 # GET
-@route("/plonesites", "plonesites", methods=["GET"])
-@route("/plonesites/<string:uid>", "plonesites", methods=["GET"])
+@route("/plonesites", "plone.jsonapi.routes.plonesites", methods=["GET"])
+@route("/plonesites/<string:uid>", "plone.jsonapi.routes.plonesites", methods=["GET"])
 def get(context, request, uid=None):
     """ Plone sites
     """
-    return get_batched("Plone Site", uid=uid, endpoint="plonesites")
+    return get_batched("Plone Site", uid=uid, endpoint="plone.jsonapi.routes.plonesites")