+++ src/plone/jsonapi/routes/zcml.py
@@ -11,17 +11,13 @@ class IConfig(Interface):
     """ Plone JSONAPI Config Interface
     """
 
-    use_advanced_query = Bool(
-        title=u"Search uses Advanced Query",
-        description=u"""
-        Use Advanced Query if installed for catalog searches
+    register_api_routes = Bool(
+        title=u"Register Plone JSON Routes",
+        description=u"""Register JSON API Routes for Plone
         """,
         required=False)
 
 
-def configDirective(_context, use_advanced_query=False):
-
-    if use_advanced_query:
-        from plone.jsonapi.routes import query
-        logger.info("*** USE ADVANCED QUERY ***")
-        query.USE_ADVANCED_QUERY = use_advanced_query
+def configDirective(_context, register_api_routes=True):
+    import plone.jsonapi.routes
+    plone.jsonapi.routes.REGISTER_API_ROUTES = register_api_routes